Fix for test when help page changed
Added currently existing default cli help page text. I think
that check for what is in help page should be either automated
instead of using hardcoded copy-pasted text or create custom
help page builder which can be reused in tests to verify if output
matched with that was built for help page.